ostatni-podpunkt-z-lab3-i-inne

+++ wsb-jpa1/src/main/java/com/jpacourse/rest/VisitController.java
@@ -17,7 +17,7 @@ public class VisitController {
     public VisitController(VisitService visitService) {this.visitService = visitService;}
 
     @GetMapping("/find-by-patientId/{patientId}")
-    public List<VisitTO> getAllVisitsByPatientId(@PathVariable final int patientId) {
+    public List<VisitTO> getAllVisitsByPatientId(@PathVariable final Long patientId) {
         return visitService.findByPatientId(patientId);
     }
 }
